pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "marco"
version = "0.1.0"
description = "AI-powered recipe generator with psychonutrition for anxiety management"
readme = "README.md"
requires-python = ">=3.10"
license = { text = "MIT" }
authors = [
    { name = "Guillaume Bailleul", email = "guillaume.bailleul@example.com" },
]
dependencies = [
    "langgraph>=0.2.0",
    "langchain>=0.3.0",
    "langchain-openai>=0.2.0",
    "langchain-anthropic>=0.2.0",
    "langchain-ollama>=0.1.0",
    "langchain-community>=0.3.0",
    "typer>=0.12.0",
    "pydantic>=2.0.0",
    "pydantic-settings>=2.0.0",
    "weasyprint>=60.0",
    "jinja2>=3.1.0",
    "python-dotenv>=1.0.0",
    "rich>=13.0.0",
]

[project.optional-dependencies]
dev = ["pytest>=8.0.0", "pytest-cov>=4.1.0", "black>=24.0.0", "ruff>=0.3.0"]

[project.scripts]
marco = "marco.cli:app"

[tool.hatch.build.targets.wheel]
packages = ["src/marco"]

[tool.black]
line-length = 100
target-version = ['py310']

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "I"]
ignore = []
